+++ src/livestreamer/plugins/streamlive.py
@@ -7,8 +7,8 @@ from livestreamer.stream import HLSStream, RTMPStream
 
 CHANNEL_URL = "http://www.mobileonline.tv/channel.php"
 
-_url_re = re.compile("http(s)?://(\w+\.)?(ilive.to|streamlive.to)/view/(?P<channel>\d+)")
-_link_re = re.compile("<p style=\"font-size:30px;\"><a href=(\S+) target=")
+_url_re = re.compile("http(s)?://(\w+\.)?(ilive.to|streamlive.to)/.*/(?P<channel>\d+)")
+_link_re = re.compile("<a href=(\S+) target=\"_blank\"")
 _schema = validate.Schema(
     validate.transform(_link_re.findall),
 )