+++ src/livestreamer_cli/utils/http_server.py
@@ -27,18 +27,38 @@ class HTTPServer(object):
         self.bound = False
 
     @property
+    def addresses(self):
+        if self.host:
+            return [self.host]
+
+        addrs = set()
+        for info in socket.getaddrinfo(socket.gethostname(), self.port,
+                                       socket.AF_INET):
+            addrs.add(info[4][0])
+
+        addrs.add("127.0.0.1")
+        return sorted(addrs)
+
+    @property
+    def urls(self):
+        for addr in self.addresses:
+            yield "http://{0}:{1}/".format(addr, self.port)
+
+    @property
     def url(self):
-        return "http://{0}:{1}/".format(self.host, self.port)
+        return next(self.urls, None)
 
     def bind(self, host="127.0.0.1", port=0):
         try:
-            self.socket.bind((host, port))
+            self.socket.bind((host or "", port))
         except socket.error as err:
             raise OSError(err)
 
         self.socket.listen(1)
         self.bound = True
         self.host, self.port = self.socket.getsockname()
+        if self.host == "0.0.0.0":
+            self.host = None
 
     def open(self, timeout=30):
         self.socket.settimeout(timeout)
@@ -55,10 +75,21 @@ class HTTPServer(object):
             raise OSError("Failed to read data from socket")
 
         req = HTTPRequest(req_data)
+        if req.command not in ("GET", "HEAD"):
+            conn.send(b"HTTP/1.1 501 Not Implemented\r\n")
+            conn.close()
+            raise OSError("Invalid request method: {0}".format(req.command))
 
         conn.send(b"HTTP/1.1 200 OK\r\n")
         conn.send(b"Server: Livestreamer\r\n")
+        conn.send(b"Content-Type: video/unknown\r\n")
         conn.send(b"\r\n")
+
+        # We don't want to send any data on HEAD requests.
+        if req.command == "HEAD":
+            conn.close()
+            raise OSError
+
         self.conn = conn
 
         return req
@@ -75,4 +106,3 @@ class HTTPServer(object):
 
         if not client_only:
             self.socket.close()
-